community/knowledge_graph_rag/requirements.txt
python-arango
pandas
tqdm
langchain
langchain-nvidia-ai-endpoints
langchain-community
networkx
nx-arangodb
